+++ changelog.md
@@ -1,1 +1,5 @@
+### 3.3.1
+- Fixed blacklisted arrow from infinity enchantment unable to pickup if bow is enchanted with infinity
+
+### 3.3.0
 - Added config to blacklist arrow from infinity enchantment